requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning & NLP
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0

# Vector Database & Embeddings
chromadb>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
sentence-transformers>=2.2.0

# Text Processing
langchain-text-splitters>=0.0.1

# Web Interface
gradio>=4.0.0

# Data Visualization (for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter Notebook Support
jupyter>=1.0.0
ipykernel>=6.20.0

# Data Loading & Processing
kagglehub>=0.2.0

# Environment Management
python-dotenv>=1.0.0

# HTTP Requests (if needed for API calls)
requests>=2.28.0

# Additional utilities
tqdm>=4.65.0
